FIX - Comparison window takes priority + combine into ankimon_sync file
